[FP-71] Feat: AddressInputSection 컴포넌트 타입 분리 (별도 파일로 이동)
